requirements.txt
django==3.0.5
djangorestframework==3.11.0
psycopg2==2.8.4
pytz==2019.3
django-cors-headers==3.2.1
django-rest-knox==4.1.0